refactor: improve folder structure fix: prevent adding to cart when not logged in (ProductCard.tsx)
