+++ client/src/components/CartItem/CartItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Minus, Plus, Trash2 } from 'lucide-react';
-import { CartItem as CartItemType } from '../types';
-import { useCart } from '../store/useCart';
+import { CartItem as CartItemType } from '../../types';
+import { useCart } from '../../store/useCart';
 
 interface CartItemProps {
   item: CartItemType;
