+++ client/src/components/CartItem/index.ts
@@ -1,0 +1,1 @@
+export {CartItem} from './CartItem';
