+++ client/src/components/Footer/Footer.tsx
@@ -1,6 +1,4 @@
-
-
-const Footer = () => {
+export const Footer = () => {
   return (
     <footer className="bg-gray-200 text-black py-4 text-center">
       <p className="text-sm">© {new Date().getFullYear()}. Todos los derechos reservados.</p>
@@ -8,5 +6,3 @@ const Footer = () => {
     </footer>
   )
 }
-
-export default Footer
