+++ client/src/components/LogoutButton/index.ts
@@ -1,0 +1,1 @@
+export {LogoutButton} from './LogoutButton';
