+++ client/src/components/ProtectedRoute/index.ts
@@ -1,0 +1,1 @@
+export {ProtectedRoute} from './ProtectedRoute';
