+++ client/src/layout/MainLayout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Footer from "../components/Footer";
-import { Navbar } from "../components/Navbar";
+import {Footer} from "../components/Footer";
+import { Navbar } from "../components/Navbar/Navbar";
 import { Outlet } from "react-router-dom";
 
 const MainLayout: React.FC = () => {
