+++ client/src/pages/home/Home.tsx
@@ -1,11 +1,8 @@
-import { ProductCard } from '../components/ProductCard';
-// import { products } from '../data/products';
-import { useProducts } from '../hooks/useProducts';
+import { ProductCard } from '../../components/ProductCard/ProductCard';
+import { useProducts } from '../../hooks/useProducts';
 
 export const Home = () => {
-
   const {products, loading, error} = useProducts();
-
   return (
     <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
       
@@ -31,7 +28,6 @@ export const Home = () => {
       {!loading && products.length > 0 && (
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
           {products.map((product,index) => (
-            // console.log("renderizando producto con ID:", product.id_p);
             <ProductCard key={`${product.id}-${index}`} product={product} />
           ))}
         </div>
