--- client/src/routes/route.js
@@ -1,18 +1,0 @@
-/**
-
-RUTAS DE LA APLICACIÓN
-
-
-/account/create-account         -> Register
-/account/login                 -> Login
-/dashboard/:username           -> Dashboard Home o Profile
-/dashboard/:username/profile   -> Perfil
-/dashboard/:username/orders    -> Historial de compras
-/dashboard/:username/wishlist  -> Lista de deseos
-/dashboard/:username/addresses -> Direcciones
-/dashboard/:username/payments  -> Métodos de pago
-/dashboard/:username/reviews   -> Reseñas
-
-
-
- */
